Added additional functionality to search
-- It should now allow a person to search for different attributes as
well as the name (BUT ONLY IN THE END SEARCH BAR!!! -- person should not
be able to search for anything but a specific location for a start
UNLESS we want to eventually add a random place generator for the start
location search bar). 

--Also added a hashmap to the attributes class, this is important if we
want users to be allowed to enter multiple spellings of certain
landmarks that wont result in the program not doing anything unless they
spell an attribute one specific way. 

--Another note: This is largely untested -- I know for a fact that
search should work when a user enters the specific name of a location
(will update to support alias' as well), but I have been unable to test
the entry of attributes due to the devgui part not being updated in
master just yet! 

--Also, need the changes that allow user to pick points from multiple
maps done on global map before I switch search to global map, because,
as it stands, nodes are still being found on local map.

--will add comments within the next couple commits: